renamed package for the orders kafka example
