+++ src/main/java/application/app/Application.java
@@ -12,9 +12,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.core.env.Environment;
 
-@ServletComponentScan({"com.ibm.hello.*"})
+@ServletComponentScan({"com.ibm.orderskafka.*"})
 @SpringBootApplication
-@ComponentScan({"com.ibm.hello.*","com.ibm.simplekafka.*", "com.ibm.cloud_garage.*", "com.ibm.health"})
+@ComponentScan({"com.ibm.hello.*","com.ibm.simplekafka.*", "com.ibm.orderskafka.*","com.ibm.cloud_garage.*", "com.ibm.health"})
 public class Application extends SpringBootServletInitializer {
     @Autowired
     Environment environment;
