+++ src/main/java/com/ibm/orderskafka/config/OrderProducerKafkaTemplate.java
@@ -1,4 +1,4 @@
-package com.ibm.hello.config.kafka.orders;
+package com.ibm.orderskafka.config;
 
 import java.util.Properties;
