+++ src/main/java/com/ibm/orderskafka/listener/AgentsInitializer.java
@@ -1,4 +1,4 @@
-package com.ibm.hello.listener.kafka.orders;
+package com.ibm.orderskafka.listener;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -12,8 +12,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import com.ibm.hello.config.kafka.orders.KafkaConfiguration;
-import com.ibm.hello.task.kafka.orders.OrderEventsAgent;
+import com.ibm.orderskafka.config.KafkaConfiguration;
+import com.ibm.orderskafka.task.OrderEventsAgent;
 
 
 /**
