+++ src/main/java/com/ibm/orderskafka/model/OrderEntity.java
@@ -1,4 +1,4 @@
-package com.ibm.hello.model.kafka.orders;
+package com.ibm.orderskafka.model;
 
 public class OrderEntity {
 	public static final String PENDING_STATUS = "pending";
