+++ src/main/java/com/ibm/orderskafka/model/OrderFactory.java
@@ -1,4 +1,4 @@
-package com.ibm.hello.model.kafka.orders;
+package com.ibm.orderskafka.model;
 
 import java.util.UUID;
